improved build to dist
index.dist.html gets copied to dist and renamed to index.html so no
manual changes to base href needed.
Also, wanted to expose the TypeScript and Spec files on the distribution
build because, this project needs to let developers see what's going on
under the hood if they use Chrome dev tools.